Check site exists before clearing it

+++ util/buildsite.py
@@ -69,11 +69,12 @@ if args.autobuild:
 def clear():
     if args.clearfirst or args.autobuild:
         print("Clearing %s directory" % OUTPUTDIR)
-        for root, dirs, files in os.walk(OUTPUTDIR):
-            for f in files:
-                os.unlink(os.path.join(root, f))
-            for d in dirs:
-                shutil.rmtree(os.path.join(root, d))
+        if os.path.isdir(OUTPUTDIR):
+            for root, dirs, files in os.walk(OUTPUTDIR):
+                for f in files:
+                    os.unlink(os.path.join(root, f))
+                for d in dirs:
+                    shutil.rmtree(os.path.join(root, d))
 
 ###################################################
 #RUN TESTS
